Move the ANNW class to a dedicated file
